BTC_StateDetector_6L512_v1/btc_direction_predictor/requirements.txt
# Core
python>=3.10
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Data & HTTP
requests>=2.31.0
pyarrow>=12.0.0  # For parquet

# ML
scikit-learn>=1.3.0
lightgbm>=4.0.0
shap>=0.42.0

# Deep Learning (choose one - using PyTorch)
torch>=2.0.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2

# Optional: Jupyter for notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Hyperparameter optimization
optuna>=3.3.0



